refactor: 🛠️ Jules05: resolve cross-agent risk and sizing conflicts
- Consolidates fragmented risk logic from `RiskEngine` into `RiskManager`.
- Standardizes `approve()` interface to return `RiskDecision` with adjusted lot sizes.
- Ported institutional ATR-based scaling and exposure limits to `RiskManager`.
- Updates `main.py` live loop to pass required context to the risk gate.
- Fixes immutable schema conflict by making `TradeSignal` mutable for sizing.
- Deletes redundant `src/trading/risk_engine.py` and updates exports.
- Updates `CONFLICT_LOG.md` with harmonization details.
