+++ src/trading/audited_risk_manager.py
@@ -9,11 +9,13 @@ License: MIT
 from __future__ import annotations
 
 import logging
-from typing import Optional
+from typing import Any, Dict, List, Optional
+
+import pandas as pd
 
 from src.core.audit_log import get_audit_logger
 from src.core.schemas import TradeSignal
-from src.trading.risk_manager import RiskManager
+from src.trading.risk_manager import RiskDecision, RiskManager
 
 logger = logging.getLogger(__name__)
 
@@ -29,24 +31,47 @@ class AuditedRiskManager(RiskManager):
         signal: TradeSignal,
         signal_id: Optional[int] = None,
         model_health: Optional[dict] = None,
-    ) -> bool:
+        market_data: Optional[pd.DataFrame] = None,
+        open_positions: Optional[List[Dict[str, Any]]] = None,
+    ) -> RiskDecision:
         """
         Run the full 8-layer risk filter cascade.
-        Returns True only if ALL layers pass.
+        Returns RiskDecision only if ALL layers pass.
         Logs the full decision chain to the audit log.
         """
+        positions = open_positions if open_positions is not None else []
+
         decision_chain = {
             "circuit_breaker": self._check_circuit_breaker(),
-            "daily_loss": self._check_daily_loss(),
-            "max_positions": self._check_max_positions(),
+            "daily_loss": self.get_daily_loss_level() < 4,
+            "max_positions": len(positions) < self.cfg.max_positions,
             "symbol_allocation": self._check_symbol_allocation(signal.symbol),
             "min_confidence": self._check_minimum_confidence(signal.confidence),
             "risk_reward": self._check_risk_reward(signal),
             "consecutive_losses": self._check_consecutive_losses(),
             "model_health": self._check_model_health(model_health),
+            "directional_exposure": self._check_directional_exposure(signal, positions),
         }
 
-        passed = all(decision_chain.values())
+        if market_data is not None:
+            decision_chain["total_notional"] = self._check_total_notional(signal, positions, market_data)
+
+        is_approved = all(decision_chain.values())
+
+        # Calculate final lot size
+        adjusted_lots = signal.lot_size
+        if is_approved and market_data is not None:
+            adjusted_lots = self.calculate_position_size(signal.symbol, market_data)
+
+        if is_approved and adjusted_lots < self.cfg.min_lot_size:
+            is_approved = False
+            decision_chain["min_lot_size"] = False
+            reason = f"Calculated lot size {adjusted_lots} below minimum"
+        elif not is_approved:
+            rejection_reasons = [k for k, v in decision_chain.items() if not v]
+            reason = f"Failed filters: {', '.join(rejection_reasons)}"
+        else:
+            reason = "Approved"
 
         # Log to Audit Trail
         try:
@@ -55,7 +80,7 @@ class AuditedRiskManager(RiskManager):
                 symbol=signal.symbol,
                 direction=signal.direction,
                 decision_chain=decision_chain,
-                passed=passed,
+                passed=is_approved,
             )
 
             # Log high-severity circuit breaker events specifically
@@ -78,20 +103,19 @@ class AuditedRiskManager(RiskManager):
         except (RuntimeError, ImportError):
             logger.debug("AuditLogger not available for risk decision logging")
 
-        if not passed:
-            rejection_reasons = [k for k, v in decision_chain.items() if not v]
-            reason_str = ", ".join(rejection_reasons)
+        if not is_approved:
             logger.warning(
-                "Signal REJECTED | %s %s | Failed: %s",
+                "Signal REJECTED | %s %s | %s",
                 signal.symbol,
                 signal.direction,
-                reason_str,
+                reason,
             )
             if self.trade_logger:
                 self.trade_logger.log_risk_event(
                     event_type="SIGNAL_REJECTED",
-                    description=f"Failed filters: {reason_str}",
+                    description=reason,
                     symbol=signal.symbol,
                     signal_id=signal_id,
                 )
-        return passed
+
+        return RiskDecision(is_approved=is_approved, reason=reason, adjusted_lot_size=adjusted_lots)
